+++ src/main/kotlin/kr/wooco/woocobe/auth/ui/filter/JwtAuthenticationFilter.kt
@@ -1,40 +1,43 @@
-package kr.wooco.woocobe.common.ui.security
+package kr.wooco.woocobe.auth.ui.filter
 
 import jakarta.servlet.FilterChain
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
-import kr.wooco.woocobe.auth.infrastructure.token.JWTProvider
+import kr.wooco.woocobe.auth.domain.usecase.ExtractTokenInput
+import kr.wooco.woocobe.auth.domain.usecase.ExtractTokenUseCase
 import org.springframework.http.HttpHeaders
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
 import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.web.filter.OncePerRequestFilter
 
 class JwtAuthenticationFilter(
-    private val jwtProvider: JWTProvider,
+    private val extractTokenUseCase: ExtractTokenUseCase,
 ) : OncePerRequestFilter() {
     override fun doFilterInternal(
         request: HttpServletRequest,
         response: HttpServletResponse,
         filterChain: FilterChain,
     ) {
-        runCatching {
-            val accessToken = extractAccessTokenInRequestHeader(request)
-            val userId = jwtProvider.extractUserId(accessToken)
-            SecurityContextHolder.getContext().apply {
-                authentication = UsernamePasswordAuthenticationToken(userId, null, emptyList())
-            }
-        }.onFailure {
-            SecurityContextHolder.clearContext()
-        }.run {
-            filterChain.doFilter(request, response)
+        val accessToken = extractAccessTokenInRequestHeader(request)
+        val results = extractTokenUseCase.execute(
+            ExtractTokenInput(
+                accessToken = accessToken,
+            ),
+        )
+
+        val authentication = UsernamePasswordAuthenticationToken(results.userId, null, emptyList())
+        SecurityContextHolder.getContext().apply {
+            this.authentication = authentication
         }
+
+        filterChain.doFilter(request, response)
     }
 
     private fun extractAccessTokenInRequestHeader(request: HttpServletRequest): String =
         request.getHeader(HttpHeaders.AUTHORIZATION)?.let {
             when (it.startsWith(BEARER_PREFIX)) {
                 true -> it.substring(BEARER_PREFIX.length)
-                else -> null
+                else -> throw RuntimeException()
             }
         } ?: throw RuntimeException()
